Refactor: Move semaphore implementation to src/semaphore.cpp
