--- h/Semaphore.h
@@ -1,7 +1,0 @@
-#ifndef OS1_SIMPLE_SEMAPHORE_H
-#define OS1_SIMPLE_SEMAPHORE_H
-
-#include "Semaphore.hpp"
-
-#endif // OS1_SIMPLE_SEMAPHORE_H
-
